+++ app.js
@@ -3,12 +3,13 @@
 // ==============================
 const
     express                 = require('express'),
+    session                 = require('express-session'),
+    mongoose                = require("mongoose"),
+    MongoStore              = require('connect-mongo')(session),
     flash                   = require('connect-flash'),
     path                    = require('path'),
     logger                  = require('morgan'),
-    cookieParser            = require('cookie-parser'),
     bodyParser              = require('body-parser'),
-    mongoose                = require("mongoose"),
     methodOverride          = require('method-override'),
     favicon                 = require('serve-favicon'),
     passport                = require('passport'),
@@ -32,6 +33,7 @@ require('./schema')._siteConfig.siteInitialization();
 //  CONFIG
 // ==============================
 // express
+app.engine('dot', require('./config/_viewEngine').__express);
 app.set('x-powered-by', false);
 app.set('view engine', 'dot');
 app.set('views', path.join(__dirname, './views'));
@@ -39,20 +41,26 @@ app.set('partials', {
     theme: path.join(__dirname, './views/theme/_partials'),
     console: path.join(__dirname, './views/console/_partials'),
 });
-app.engine('dot', require('./config/_viewEngine').__express);
+app.use(express.static(path.join(__dirname, './public'), {
+    setHeaders: (res, path, stat) => res.set('x-robots-tag', 'none'),
+}));
+app.use(session({
+    secret: process.env.SECRET,
+    saveUninitialized: false,
+    resave: false,
+    // cookie: { secure: true },
+    store: new MongoStore({
+        mongooseConnection: mongoose.connection,
+        autoRemove: 'native',
+    }),
+}));
 
 
 // dependencies
 if (process.env.ENV === 'dev') app.use(logger('dev'));
-app.use(express.static(path.join(__dirname, './public'), {
-    setHeaders: (res, path, stat) => {
-        res.set('x-robots-tag', 'none');
-    }
-}));
 app.use(methodOverride('_method'));
-app.use(bodyParser.urlencoded({extended: true}));
+app.use(bodyParser.urlencoded({ extended: true }));
 app.use(bodyParser.json());
-app.use(cookieParser());
 app.use(flash());
 //app.use(favicon(path.join(__dirname, 'public', 'favicon.ico')));
 require('./config/passport')(app, passport);